+++ helpers/utils.js
@@ -4,19 +4,28 @@ import { RE_IS_SCHEDULED } from './dateTime'
 import { RE_ARROW_DATES_G } from './regex'
 
 /**
- * Test whether a task is open or not (type: 'open' or 'checklist')
+ * Test whether a para is open or not (type: 'open' or 'checklist')
  * @param {Paragraph} t - the paragraph/task to check
  * @returns {boolean} true if open, false if any other status/type
  */
 export const isOpen = (t: TParagraph): boolean => t.type === 'open' || t.type === 'checklist'
 
 /**
- * Test whether a task is open or not (type: 'open' or 'checklist') and doesn't have a scheduled date
+ * Test whether a para is open or not (type: 'open' or 'checklist') and doesn't have a scheduled date
  * @param {Paragraph} t - the paragraph/task to check
  * @returns {boolean} true if open, false if any other status/type
  */
 export function isOpenNotScheduled(t: TParagraph): boolean {
   return (t.type === 'open' || t.type === 'checklist') && !hasScheduledDate(t.content)
+}
+
+/**
+ * Test whether a para is open or not (type: 'open') and doesn't have a scheduled date
+ * @param {Paragraph} t - the paragraph/task to check
+ * @returns {boolean} true if open, false if any other status/type
+ */
+export function isOpenTaskNotScheduled(t: TParagraph): boolean {
+  return (t.type === 'open') && !hasScheduledDate(t.content)
 }
 
 /**